# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Next.js API Documentation"
  },
...
  "paths": {
...
    "/api/{lang}/news": {
      "post": {
        "security": [{ "cookieAuth": [] }],
        "tags": ["Home - News"],
        "summary": "Create news",
        "description": "Send an empty request body",
        "parameters": [
          {
            "name": "lang",
            "in": "path",
            "required": true,
            "description": "Language code (e.g., 'uk', 'en')",
            "schema": {
              "type": "string",
              "enum": ["uk", "en"],
              "example": "en"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Object with news content",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "res": {
                      "type": "object",
                      "$ref": "#/components/schemas/ResponseOfCreateCard"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/api/{lang}/news/{id}": {
      "delete": {
        "security": [{ "cookieAuth": [] }],
        "tags": ["Home - News"],
        "summary": "Delete news",
        "description": "Send an empty request body",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "News ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Object with news content",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "News deleted"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "400":{
            "description": "Bad request"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "_id": {
        "type": "string"
      },
      "ResponseOfCreateCard": {
        "type": "object",
        "properties": {
          "_id": {
            "$ref": "#/components/schemas/_id"
          },
          "header": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "url": {
            "type": "string",
            "example": "url"
          }
        }
      },
...
    }
  }
}